html/canvas/offscreen/2d.text.measure.width.space.worker-expected.txt

This is a testharness.js-based test.
FAIL Space characters are converted to U+0020 and collapsed (per CSS) for OffscreenCanvas assert_equals: ctx.measureText('A  B').width === 150 (got 200[number], expected 150[number]) expected 150 but got 200
Harness: the test ran to completion.
